+++ gasmodel/Pact/Core/GasModel/Serialization.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE TypeApplications #-}
 module Pact.Core.GasModel.Serialization
   (benchmarks) where
