+++ pact/Pact/Core/Gas.hs
@@ -5,7 +5,9 @@
 module Pact.Core.Gas
  ( module Pact.Core.Gas.Types
  , module Pact.Core.Gas.Utils
+ , module Pact.Core.Gas.TableGasModel
  ) where
 
 import Pact.Core.Gas.Types
 import Pact.Core.Gas.Utils
+import Pact.Core.Gas.TableGasModel
